Adds staking operations to CLI and JS SDK (#307)
* feat: Adds staking operations to CLI and JS SDK

This commit introduces staking capabilities for both the `genlayer` CLI and the `genlayer-js` SDK.

It includes commands for validators and delegators to manage their stakes on the testnet-asimov network such as joining, exiting, and claiming rewards.
Also provides commands to manage validator identity.

The SDK provides functions to get staking information, perform stake actions, and helper utilities.

* Feat: Adds account import functionality

Adds the ability to import an account using a private key
via the genlayer CLI.

Includes options for specifying the private key and output path.